#36 extract getSuggestions with modifications for function call
